requirements.txt
Django~=4.1.0
djangorestframework==3.13.1